+++ src/js/constants.js
@@ -1,0 +1,16 @@
+export const YOUTUBE_SEARCH_ACTION = {
+  UPDATE_SEARCH_KEYWORD: 'UPDATE_SEARCH_KEYWORD',
+  UPDATE_SEARCH_RESULT: 'UPDATE_SEARCH_RESULT',
+  UPDATE_SEARCH_LOADING_STATUS: 'UPDATE_SEARCH_LOADING_STATUS',
+};
+
+export const ERROR_MESSAGE = {
+  EMPTY_SEARCH_KEYWORD: '검색어를 입력해주세요.',
+  MAX_SAVE_VIDEO: '동영상은 최대 100개까지 저장할 수 있습니다.',
+};
+
+export const YOUTUBE_SETTING = {
+  URI: 'https://www.googleapis.com/youtube/v3/search',
+  MAX_VIDEO_NUMBER: 10,
+  MAX_SAVE_NUMBER: 100,
+};
