--- src/js/constants/action.js
@@ -1,5 +1,0 @@
-export const YOUTUBE_SEARCH_ACTION = {
-  UPDATE_SEARCH_KEYWORD: 'UPDATE_SEARCH_KEYWORD',
-  UPDATE_SEARCH_RESULT: 'UPDATE_SEARCH_RESULT',
-  UPDATE_SEARCH_LOADING_STATUS: 'UPDATE_SEARCH_LOADING_STATUS',
-};
