--- src/js/constants/message.js
@@ -1,4 +1,0 @@
-export const ERROR_MESSAGE = {
-  EMPTY_SEARCH_KEYWORD: '검색어를 입력해주세요.',
-  MAX_SAVE_VIDEO: '동영상은 최대 100개까지 저장할 수 있습니다.',
-};
