--- src/js/constants/setting.js
@@ -1,5 +1,0 @@
-export const YOUTUBE_SETTING = {
-  URI: 'https://www.googleapis.com/youtube/v3/search',
-  MAX_VIDEO_NUMBER: 10,
-  MAX_SAVE_NUMBER: 100,
-};
